shopbot/requirements.txt
# ═══════════════════════════════════════════════════
# ShopBot — Dépendances Python
# Installation : pip install -r requirements.txt
# ═══════════════════════════════════════════════════

# LangChain core
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-mistralai>=0.1.0

# Base vectorielle
chromadb>=0.5.0

# Chargement de documents PDF
pypdf>=4.0.0

# Recherche web
tavily-python>=0.3.0

# API REST
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Variables d'environnement
python-dotenv>=1.0.0

# Validation de données
pydantic>=2.0.0
